refactoring(Auth) create Auth servies to extract same logic
