+++ server/src/services/email.service.ts
@@ -1,4 +1,90 @@
 import { transporter, createEmailTemplate } from '../config/mailer.js';
+import config from '../config/config.js';
+import { IUser } from '../types/user.types.js';
+
+/**
+ * Serwis do obsługi wysyłania wiadomości email
+ */
+export class EmailService {
+  /**
+   * Wysyła email powitalny do nowego użytkownika
+   * @param user - Obiekt użytkownika
+   */
+  async sendWelcomeEmail(user: IUser): Promise<void> {
+    const subject = 'Witamy w CodeLinesJS!';
+    const html = createEmailTemplate({
+      title: 'Witaj w CodeLinesJS!',
+      preheader: 'Dziękujemy za dołączenie do naszej społeczności.',
+      content: `
+        <p>Cześć ${user.username}!</p>
+        <p>Cieszymy się, że dołączyłeś/aś do naszej społeczności. Jesteśmy tutaj, aby pomóc Ci rozwijać umiejętności programowania w JavaScript.</p>
+        <p>Możesz zalogować się do swojego konta używając adresu email: ${user.email}</p>
+        <p>Pozdrawiamy,<br/>Zespół CodeLinesJS</p>
+      `
+    });
+
+    await transporter.sendMail({
+      from: config.email.from,
+      to: user.email,
+      subject,
+      html
+    });
+  }
+
+  /**
+   * Wysyła email z instrukcjami resetowania hasła
+   * @param user - Obiekt użytkownika
+   * @param resetUrl - Link do resetowania hasła
+   */
+  async sendPasswordResetEmail(user: IUser, resetUrl: string): Promise<void> {
+    const subject = 'Resetowanie hasła w CodeLinesJS';
+    const html = createEmailTemplate({
+      title: 'Resetowanie hasła',
+      preheader: 'Instrukcje dotyczące resetowania hasła w CodeLinesJS.',
+      content: `
+        <p>Cześć ${user.username}!</p>
+        <p>Otrzymaliśmy prośbę o resetowanie Twojego hasła. Jeśli to nie Ty, zignoruj tę wiadomość.</p>
+        <p>Aby zresetować hasło, kliknij w poniższy link (ważny przez 1 godzinę):</p>
+        <p><a href="${resetUrl}" style="padding: 10px 15px; background-color: #4CAF50; color: white; text-decoration: none; border-radius: 5px;">Resetuj hasło</a></p>
+        <p>Jeśli link nie działa, skopiuj i wklej poniższy URL do przeglądarki:</p>
+        <p>${resetUrl}</p>
+        <p>Pozdrawiamy,<br/>Zespół CodeLinesJS</p>
+      `
+    });
+
+    await transporter.sendMail({
+      from: config.email.from,
+      to: user.email,
+      subject,
+      html
+    });
+  }
+
+  /**
+   * Wysyła email potwierdzający zmianę hasła
+   * @param user - Obiekt użytkownika
+   */
+  async sendPasswordChangedEmail(user: IUser): Promise<void> {
+    const subject = 'Potwierdzenie zmiany hasła w CodeLinesJS';
+    const html = createEmailTemplate({
+      title: 'Hasło zostało zmienione',
+      preheader: 'Potwierdzenie zmiany hasła w CodeLinesJS.',
+      content: `
+        <p>Cześć ${user.username}!</p>
+        <p>Twoje hasło zostało pomyślnie zmienione.</p>
+        <p>Jeśli to nie Ty zmieniłeś/aś hasło, natychmiast skontaktuj się z nami.</p>
+        <p>Pozdrawiamy,<br/>Zespół CodeLinesJS</p>
+      `
+    });
+
+    await transporter.sendMail({
+      from: config.email.from,
+      to: user.email,
+      subject,
+      html
+    });
+  }
+}
 
 export const sendBugReportConfirmation = async (email: string, title: string) => {
   try {
